break up functions.h into another header that has non_wx std functions to include elsewhere, particularly my catch2 tests
