+++ src/constants/cistem_numbers.h
@@ -1,6 +1,8 @@
 #ifndef _SRC_CONSTANTS_CISTEM_NUMBERS_H_
 #define _SRC_CONSTANTS_CISTEM_NUMBERS_H_
 
+#include <complex>
+
 #if __cplusplus > 201703L
 #include <numbers>
 using namespace std::numbers;
@@ -41,6 +43,8 @@ inline constexpr _Tp sqrt3_v = static_cast<_Tp>(1.732050807568877293527446341505
 
 template <typename _Tp>
 inline constexpr _Tp inv_sqrt3_v = _Tp(1) / sqrt3_v<_Tp>;
+
+constexpr std::complex<float> I(0.0, 1.0);
 
 #endif
